Feature/edit chair 30 dic 2019

+++ src/app/components/main/chairs-list/chairs-list.component.ts
@@ -3,7 +3,9 @@ import { AlertModel } from '../../../models/alert.model';
 import { ChairModel } from '../../../models/chair.model';
 import { CrudService } from '../../../services/crud.service';
 import { MatDialog } from '@angular/material';
-import { AddChairComponent } from 'src/app/components/main/add-chair/add-chair.component'
+import { AddChairComponent } from 'src/app/components/main/add-chair/add-chair.component';
+import { ConfirmationModalComponent } from '../../shared/confirmation-modal/confirmation-modal.component';
+
 @Component({
   selector: 'app-chairs-list',
   templateUrl: './chairs-list.component.html',
@@ -78,19 +80,62 @@ export class ChairsListComponent implements OnInit {
     });
   }
 
+  editChair(chair: ChairModel) {
+    console.log(chair);
+    const dialogRef = this.dialog.open(AddChairComponent, {
+      width: '600px',
+      data: {
+        id: chair.id,
+        description: chair.description,
+        owner: chair.owner,
+        location: chair.location
+      }
+    });
+    dialogRef.componentInstance.buttonText = 'Edit chair';
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result !== undefined && result !== false) {
+        this.crudService.updateDocument(this.collectionName, chair.id, result)
+        .then(() => {
+          this.alert = [{
+            type: 'success',
+            message: 'Chair edited successfully!'
+          }];
+        })
+        .catch(error => {
+          this.alert = [{
+            type: 'danger',
+            message: 'Something went wrong: ' + error
+          }];
+        });
+      }
+    });
+  }
+
   deleteChair(chairId) {
-    this.crudService.deleteDocument(this.collectionName, chairId)
-      .then(() => {
-        this.alert = [{
-          type: 'success',
-          message: 'Chair deleted successfully!'
-        }];
-      })
-      .catch(error => {
-        this.alert = [{
-          type: 'danger',
-          message: 'Something went wrong: ' + error
-        }];
-      });
+    const dialogRef = this.dialog.open(ConfirmationModalComponent, {
+      width: '600px',
+      data: {
+        buttonText: 'Delete',
+        modalText: 'Are you sure that you want to delete chair?',
+        title: 'Delete chair'
+      }
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result !== undefined && result === true) {
+        this.crudService.deleteDocument(this.collectionName, chairId)
+          .then(() => {
+            this.alert = [{
+              type: 'success',
+              message: 'Chair deleted successfully!'
+            }];
+          })
+          .catch(error => {
+            this.alert = [{
+              type: 'danger',
+              message: 'Something went wrong: ' + error
+            }];
+          });
+      }
+    });
   }
 }
